calculadora_cientifica/backend/requirements.txt
Flask
python-dotenv
dot